[refactor] Move enums to separate data/enums directory
