+++ lib/data/models/recipient_form.dart
@@ -1,6 +1,6 @@
 import 'package:danawallet/data/models/recipient_form_filled.dart';
 import 'package:danawallet/data/models/recommended_fee_model.dart';
-import 'package:danawallet/data/models/selected_fee.dart';
+import 'package:danawallet/data/enums/selected_fee.dart';
 import 'package:danawallet/generated/rust/api/structs.dart';
 
 // this singleton class contains all filled data during the send flow
